+++ apps/e2e/tests/backend/endpoints/api/v1/users.test.ts
@@ -1,6 +1,6 @@
 import { generateSecureRandomString } from "@stackframe/stack-shared/dist/utils/crypto";
 import { describe } from "vitest";
-import { it } from "../../../../helpers";
+import { createMailbox, it } from "../../../../helpers";
 import { Auth, InternalProjectKeys, Project, backendContext, niceBackendFetch } from "../../../backend-helpers";
 
 describe("without project access", () => {
@@ -85,6 +85,7 @@ describe("with client access", () => {
           "id": "<stripped UUID>",
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -116,6 +117,7 @@ describe("with client access", () => {
           "id": "<stripped UUID>",
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -192,6 +194,7 @@ describe("with client access", () => {
           "id": "<stripped UUID>",
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -222,6 +225,7 @@ describe("with client access", () => {
           "id": "<stripped UUID>",
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -362,6 +366,7 @@ describe("with client access", () => {
           "id": "<stripped UUID>",
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -392,6 +397,7 @@ describe("with client access", () => {
           "id": "<stripped UUID>",
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -520,6 +526,7 @@ describe("with client access", () => {
           "id": "<stripped UUID>",
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_verified": true,
           "profile_image_url": null,
@@ -628,6 +635,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
@@ -665,6 +673,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
@@ -706,7 +715,7 @@ describe("with server access", () => {
       NiceResponse {
         "status": 200,
         "body": {
-          "is_paginated": false,
+          "is_paginated": true,
           "items": [
             {
               "auth_with_email": true,
@@ -718,6 +727,7 @@ describe("with server access", () => {
               "last_active_at_millis": <stripped field 'last_active_at_millis'>,
               "oauth_providers": [],
               "otp_auth_enabled": true,
+              "passkey_auth_enabled": false,
               "primary_email": "<stripped UUID>@stack-generated.example.com",
               "primary_email_auth_enabled": true,
               "primary_email_verified": true,
@@ -729,10 +739,38 @@ describe("with server access", () => {
               "signed_up_at_millis": <stripped field 'signed_up_at_millis'>,
             },
           ],
+          "pagination": { "next_cursor": null },
         },
         "headers": Headers { <some fields may have been hidden> },
       }
     `);
+  });
+
+  it("list next cursor", async ({ expect }) => {
+    await Project.createAndSwitch({ config: { magic_link_enabled: true } });
+    for (let i = 0; i < 5; i++) {
+      backendContext.set({ mailbox: createMailbox() });
+      await Auth.Otp.signIn();
+    }
+    const allResponse = await niceBackendFetch("/api/v1/users", {
+      accessType: "server",
+    });
+
+    const response1 = await niceBackendFetch("/api/v1/users?limit=2", {
+      accessType: "server",
+    });
+    expect(response1.body.pagination.next_cursor).toBeDefined();
+
+    const response2 = await niceBackendFetch(`/api/v1/users?limit=3&cursor=${response1.body.pagination.next_cursor}`, {
+      accessType: "server",
+    });
+    expect(response2.body.pagination.next_cursor).toBeDefined();
+
+    // check if response 1 + response 2 = allResponse
+    expect(response1.body.items.length + response2.body.items.length).toEqual(allResponse.body.items.length);
+    const allUserIds = new Set(allResponse.body.items.map((user: any) => user.id));
+    const concatenatedUserIds = new Set([...response1.body.items.map((user: any) => user.id), ...response2.body.items.map((user: any) => user.id)]);
+    expect(concatenatedUserIds).toEqual(allUserIds);
   });
 
   it("should be able to read a user", async ({ expect }) => {
@@ -760,6 +798,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
@@ -795,6 +834,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": false,
+          "passkey_auth_enabled": false,
           "primary_email": null,
           "primary_email_auth_enabled": false,
           "primary_email_verified": false,
@@ -834,6 +874,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": false,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": false,
@@ -873,6 +914,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": false,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": false,
@@ -939,6 +981,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": false,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": false,
@@ -996,6 +1039,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": false,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": false,
           "primary_email_verified": false,
@@ -1032,6 +1076,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": false,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": false,
@@ -1083,6 +1128,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": false,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": false,
@@ -1116,6 +1162,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": false,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": false,
           "primary_email_verified": false,
@@ -1173,6 +1220,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
@@ -1202,6 +1250,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
@@ -1240,6 +1289,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
@@ -1278,6 +1328,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
@@ -1365,6 +1416,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "<stripped UUID>@stack-generated.example.com",
           "primary_email_auth_enabled": true,
           "primary_email_verified": true,
@@ -1414,6 +1466,7 @@ describe("with server access", () => {
           "last_active_at_millis": <stripped field 'last_active_at_millis'>,
           "oauth_providers": [],
           "otp_auth_enabled": true,
+          "passkey_auth_enabled": false,
           "primary_email": "new-primary-email@example.com",
           "primary_email_auth_enabled": false,
           "primary_email_verified": true,