+++ packages/stack/src/components-page/account-settings.tsx
@@ -7,7 +7,7 @@ import { yupObject, yupString } from '@stackframe/stack-shared/dist/schema-field
 import { generateRandomValues } from '@stackframe/stack-shared/dist/utils/crypto';
 import { throwErr } from '@stackframe/stack-shared/dist/utils/errors';
 import { runAsynchronously, runAsynchronouslyWithAlert } from '@stackframe/stack-shared/dist/utils/promises';
-import { Accordion, AccordionContent, AccordionItem, AccordionTrigger, ActionCell, Badge, Button, Input, Label, PasswordInput, Separator, Switch, Table, TableBody, TableCell, TableHead, TableHeader, TableRow, Typography } from '@stackframe/stack-ui';
+import { Accordion, AccordionContent, AccordionItem, AccordionTrigger, ActionCell, Badge, Button, Input, Label, PasswordInput, Separator, Table, TableBody, TableCell, TableHead, TableHeader, TableRow, Typography } from '@stackframe/stack-ui';
 import { CirclePlus, Contact, Edit, LucideIcon, Settings, ShieldCheck } from 'lucide-react';
 import { useRouter } from "next/navigation";
 import { TOTPController, createTOTPKeyURI } from "oslo/otp";
@@ -329,23 +329,111 @@ function EmailsAndAuthPage() {
   const passwordSection = usePasswordSection();
   const mfaSection = useMfaSection();
   const otpSection = useOtpSection();
+  const passkeySection = usePasskeySection();
 
   return (
     <PageLayout>
       <EmailsSection/>
       {passwordSection}
+      {passkeySection}
       {otpSection}
       {mfaSection}
     </PageLayout>
   );
 }
 
+
+function usePasskeySection() {
+  const { t } = useTranslation();
+  const user = useUser({ or: "throw" });
+  const stackApp = useStackApp();
+  const project = stackApp.useProject();
+  const contactChannels = user.useContactChannels();
+
+
+  // passkey is enabled if there is a passkey
+  const hasPasskey = user.passkeyAuthEnabled;
+
+  const isLastAuth = user.passkeyAuthEnabled && !user.hasPassword && user.oauthProviders.length === 0 && !user.otpAuthEnabled;
+  const [showConfirmationModal, setShowConfirmationModal] = useState(false);
+
+  // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+  const hasValidEmail = contactChannels.filter(x => x.type === 'email' && x.isVerified && x.usedForAuth).length > 0;
+
+  if (!project.config.passkeyEnabled) {
+    return null;
+  }
+
+  const handleDeletePasskey = async () => {
+    await user.update({ passkeyAuthEnabled: false });
+    setShowConfirmationModal(false);
+  };
+
+
+  const handleAddNewPasskey = async () => {
+    await user.registerPasskey();
+  };
+
+  return (
+    <>
+      <Section title={t("Passkey")} description={hasPasskey ? t("Passkey registered") : t("Register a passkey")}>
+        <div className='flex md:justify-end gap-2'>
+          {!hasValidEmail && (
+            <Typography variant='secondary' type='label'>{t("To enable Passkey sign-in, please add a verified email and set it as your sign-in email.")}</Typography>
+          )}
+          {hasValidEmail && hasPasskey && isLastAuth && (
+            <Typography variant='secondary' type='label'>{t("Passkey sign-in is enabled and cannot be disabled as it is currently the only sign-in method")}</Typography>
+          )}
+          {!hasPasskey && hasValidEmail && (
+            <div>
+              <Button onClick={handleAddNewPasskey} variant='secondary'>{t("Add new passkey")}</Button>
+            </div>
+          )}
+          {hasValidEmail && hasPasskey && !isLastAuth && !showConfirmationModal && (
+            <Button
+              variant='secondary'
+              onClick={() => setShowConfirmationModal(true)}
+            >
+              {t("Delete Passkey")}
+            </Button>
+          )}
+          {hasValidEmail && hasPasskey && !isLastAuth && showConfirmationModal && (
+            <div className='flex flex-col gap-2'>
+              <Typography variant='destructive'>
+                {t("Are you sure you want to disable Passkey sign-in? You will not be able to sign in with your passkey anymore.")}
+              </Typography>
+              <div className='flex gap-2'>
+                <Button
+                  variant='destructive'
+                  onClick={handleDeletePasskey}
+                >
+                  {t("Disable")}
+                </Button>
+                <Button
+                  variant='secondary'
+                  onClick={() => setShowConfirmationModal(false)}
+                >
+                  {t("Cancel")}
+                </Button>
+              </div>
+            </div>
+          )}
+        </div>
+      </Section>
+
+
+    </>
+
+  );
+}
+
+
 function useOtpSection() {
   const { t } = useTranslation();
   const user = useUser({ or: "throw" });
   const project = useStackApp().useProject();
   const contactChannels = user.useContactChannels();
-  const isLastAuth = user.otpAuthEnabled && !user.hasPassword && user.oauthProviders.length === 0;
+  const isLastAuth = user.otpAuthEnabled && !user.hasPassword && user.oauthProviders.length === 0 && !user.passkeyAuthEnabled;
   const [disabling, setDisabling] = useState(false);
 
   // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition