+++ examples/middleware/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @stackframe/example-middleware-demo
 
+## 2.6.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.6.16
+
+## 2.6.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.6.15
+
+## 2.6.14
+
+### Patch Changes
+
+- @stackframe/stack@2.6.14
+
+## 2.6.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.6.13
+
+## 2.6.12
+
+### Patch Changes
+
+- Updated account settings page
+- Updated dependencies
+  - @stackframe/stack@2.6.12
+
 ## 2.6.11
 
 ### Patch Changes