+++ packages/stack-sc/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/stack-sc
 
+## 2.6.16
+
+## 2.6.15
+
+## 2.6.14
+
+## 2.6.13
+
+## 2.6.12
+
+### Patch Changes
+
+- Updated account settings page
+
 ## 2.6.11
 
 ### Patch Changes